Move code instantiating authentication daemon to a shared place.
