+++ config/initializers/config.rb
@@ -20,11 +20,43 @@
 # https://github.com/alces-software/alces-storage-manager
 #==============================================================================
 
+require 'daemon_client'
+
 module AlcesStorageManager
   class << self
     def config
       @config ||= YAML.load_file(Rails.root.join("config", "storagemanager.yml"))
     end
+
+    def authentication_daemon
+      DaemonClient::Connection.new(connection_opts)
+    end
+    
+    private
+    
+    def connection_opts
+      auth_config = AlcesStorageManager::config[:auth].dup
+      {
+        timeout: 5,
+        ssl_config: auth_config.delete(:ssl) ? ssl_config : nil
+      }.merge(auth_config)
+    end
+  
+    def ssl_config
+      @my_ssl ||= Class.new do
+        include Alces::Tools::SSLConfigurator
+        def ssl
+          ssl_opts = AlcesStorageManager::config[:ssl].dup
+          Alces::Tools::SSLConfigurator::Configuration.new(
+            root: ssl_opts[:root],
+            certificate: ssl_opts[:certificate],
+            key: ssl_opts[:key],
+            ca: ssl_opts[:ca]
+          )
+        end
+      end.new().ssl_config
+    end
+    
   end
 end
